backend/requirements.txt
fastapi>=0.110.0
uvicorn[standard]==0.24.0
sqlalchemy==2.0.23
psycopg2-binary==2.9.9
alembic==1.12.1
pydantic==2.5.0
pydantic-settings==2.1.0
python-multipart==0.0.6
httpx==0.25.2
pillow==10.1.0
python-dotenv==1.0.0
pdf2image==1.16.3
beautifulsoup4==4.12.2
requests==2.31.0

google-genai
